Better credential renewal handling
Commit 47f653e was incorrect and credential renewal is necessary,
especially for the dynamic AWS backend. Re-introduce renewal logic but
handle failures more correctly by propagating them up to the main
function by a channel so the enter process can be terminated. Previously
this used a Fatalf in a goroutine which has unpredictable interactions
with the scheduler and was resulting in the process hanging in an
intermediate state when credential renewal failed.

Also adds better logging for credential renewal.